test: move shared tests
Part of the test cleanup.
