+++ test/unit/shared/types/preferences.spec.ts
