+++ test/unit/shared/utils/git-providers.spec.ts
