+++ test/unit/shared/utils/parse-package-param.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { parsePackageParam } from '../../shared/utils/parse-package-param'
+import { parsePackageParam } from '../../../../shared/utils/parse-package-param'
 
 describe('parsePackageParam', () => {
   describe('unscoped packages', () => {
