+++ test/unit/shared/utils/spdx.spec.ts
